requirements.txt
Flask
pytesseract
pdf2docx
Pillow
